requirements.txt
requests==2.19.1
mysql-connector